Move reduxShape under utils. Refs #192
